\name{DirectStandardisation-package}
\alias{DirectStandardisation-package}
\alias{DirectStandardisation}
\docType{package}
\title{
\packageTitle{DirectStandardisation}
}
\description{
Adjusted Means and Proportions by Direct Standardisation
% Directly standardised means and proportions.
}
\details{
Calculates adjusted means and proportions of a variable by groups defined by another variable, standardised to the structure of the adjustment variables in the whole dataset using direct standardisation.

\packageIndices{DirectStandardisation}
}
\author{
\packageAuthor{DirectStandardisation}

Maintainer: \packageMaintainer{DirectStandardisation}
}

\keyword{ direct standardisation }
\keyword{ direct standardization }
